+++ flashcard/src/ui/interfaces/card.py
@@ -1,5 +1,4 @@
 from tkinter import ttk
-from tkinter.font import Font
 from ui.interfaces.template import TkinterGUITemplate
 
 
@@ -8,22 +7,18 @@ class TkinterGUICard(TkinterGUITemplate):
         super().__init__(window, service, handler)
 
         self.__card = None
-
-        # Fontit
-        self.__font_sentence = Font(size=12, weight="bold")
-        self.__font_highlight = Font(size=12, underline=1, weight="bold")
-        self.__font_reading = Font(size=11)
-        self.__font_translation = Font(size=11)
+        self.__card_turned = False
 
         self._service.generate_pack_random_order()
-        self.__initialize()
+        self._initialize()
 
     def _reinitialize(self):
         self._window.grid_rowconfigure(0, weight=10)
         self._window.grid_rowconfigure(0, weight=1)
         self._window.grid_columnconfigure(0, weight=1)
 
-    def __initialize(self):
+    def _initialize(self):
+        super()._initialize()
         self._reinitialize()
 
         frame = ttk.Frame(master=self._window, padding=(20, 20))
@@ -32,16 +27,16 @@ class TkinterGUICard(TkinterGUITemplate):
         sentenceframe = ttk.Frame(frame)
         self._add_elem(sentenceframe, 0, 0, name="frame_sentence")
         self._add_elem(ttk.Label(
-            sentenceframe, text="[SENTENCE", font=self.__font_sentence), 0, 0, name="label_sentence1", sticky="W")
-        self._add_elem(ttk.Label(sentenceframe, text="**HIGHLIGHT**", font=self.__font_highlight,
+            sentenceframe, text="[SENTENCE", font=self._handler.font_sentence), 0, 0, name="label_sentence1", sticky="W")
+        self._add_elem(ttk.Label(sentenceframe, text="**HIGHLIGHT**", font=self._handler.font_sentence_highlight,
                        foreground="#224894"), 1, 0, name="label_sentence2", sticky="W")
         self._add_elem(ttk.Label(sentenceframe, text="SENTENCE]",
-                       font=self.__font_sentence), 2, 0, name="label_sentence3", sticky="W")
+                       font=self._handler.font_sentence), 2, 0, name="label_sentence3", sticky="W")
 
         self._add_elem(ttk.Label(
-            frame, text="[READING]", font=self.__font_reading), 0, 1, name="label_reading", sticky="W")
+            frame, text="[READING]", font=self._handler.font_reading), 0, 1, name="label_reading", sticky="W")
         self._add_elem(ttk.Label(
-            frame, text="[TRANSLATION]", font=self.__font_translation), 0, 2, name="label_translation", sticky="W")
+            frame, text="[TRANSLATION]", font=self._handler.font_translation), 0, 2, name="label_translation", sticky="W")
 
         optionframe = ttk.Frame(self._window, padding=(20, 20))
         optionframe.grid_columnconfigure(0, weight=1)
@@ -54,36 +49,45 @@ class TkinterGUICard(TkinterGUITemplate):
         self._add_elem(ttk.Button(optionframe, text="Lopeta",
                                   command=self._exit), 0, 0, sticky="NE")
 
-        self.__next()
+        self.__load_text()
+
+    def __load_text(self):
+        if self.__card is None:
+            self.__next()
+
+        # Jaa lause **-merkkien perusteella
+        sentences = self.__card.sentence.split("**")
+        if len(sentences) != 3:
+            sentences = [self.__card.sentence, "", ""]
+
+        for i in range(3):
+            text = ""
+            if i < len(sentences):
+                text = sentences[i]
+            self._get_elem(f"label_sentence{i+1}").configure(text=text)
+
+
+        # Tee toimintoja riippuen onko kortti käännetty vai ei
+        if self.__card_turned:
+            self._get_elem("label_reading").configure(text=self.__card.reading)
+            self._get_elem("label_translation").configure(
+                text=self.__card.translation)
+
+            self._get_elem("button_next").configure(text="Seuraava")
+        else:
+            self._get_elem("label_reading").configure(text="")
+            self._get_elem("label_translation").configure(text="")
+            self._get_elem("button_next").configure(text="Näytä")
 
     def __next(self):
-        if self.__card is None:
+        if self.__card_turned:
             self.__card = self._service.get_next_card()
             if self.__card is None:
                 self._exit()
                 return
 
-            # Jaa lause **-merkkien perusteella
-            sentences = self.__card.sentence.split("**")
-            if len(sentences) != 3:
-                sentences = [self.__card.sentence, "", ""]
-
-            for i in range(3):
-                text = ""
-                if i < len(sentences):
-                    text = sentences[i]
-                self._get_elem(f"label_sentence{i+1}").configure(text=text)
-
-            # self._get_elem("label_sentence").configure(text=self.__card.sentence)
-            self._get_elem("label_reading").configure(text="")
-            self._get_elem("label_translation").configure(text="")
-
-            self._get_elem("button_next").configure(text="Näytä")
-
+            self.__card_turned = False
         else:
-            self._get_elem("label_reading").configure(text=self.__card.reading)
-            self._get_elem("label_translation").configure(
-                text=self.__card.translation)
+            self.__card_turned = True
 
-            self.__card = None
-            self._get_elem("button_next").configure(text="Seuraava")
+        self.__load_text()